Update docstrings. Update notebook.
